testing/web-platform/tests/css/compositing/parsing/background-blend-mode-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Compositing and Blending Level 1: getComputedStyle().backgroundBlendMode</title>
<link rel="help" href="https://drafts.fxtf.org/compositing-1/#propdef-background-blend-mode">
<meta name="assert" content="background-blend-mode computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-blend-mode", "normal");
test_computed_value("background-blend-mode", "multiply");
test_computed_value("background-blend-mode", "screen");
test_computed_value("background-blend-mode", "overlay");
test_computed_value("background-blend-mode", "darken");
test_computed_value("background-blend-mode", "lighten");
test_computed_value("background-blend-mode", "color-dodge");
test_computed_value("background-blend-mode", "color-burn");
test_computed_value("background-blend-mode", "hard-light");
test_computed_value("background-blend-mode", "soft-light");
test_computed_value("background-blend-mode", "difference");
test_computed_value("background-blend-mode", "exclusion");
test_computed_value("background-blend-mode", "hue");
test_computed_value("background-blend-mode", "saturation");
test_computed_value("background-blend-mode", "color");
test_computed_value("background-blend-mode", "luminosity");

// per spec, excess values should be dropped.
// see https://drafts.fxtf.org/compositing-1/#background-blend-mode
// and https://drafts.csswg.org/css-backgrounds-3/#layering
// > The lists are matched up from the first value: excess values at the end are not used.
test_computed_value("background-blend-mode", "normal, luminosity", "normal");
test_computed_value("background-blend-mode", "screen, overlay", "screen");
test_computed_value("background-blend-mode", "color, saturation", "color");
</script>
</body>
</html>
